multilib: use cmodel=medany when multilib is enabled
This change mitigates a number of linker errors due to unsupported
relocation types when multilib is enabled.

A typical error would be of the form:

```
relocation truncated to fit: R_RISCV_HI20 against `.LC0'
```

There are a number of similar issues filed, but that
example comes from riscv/riscv-gnu-toolchain#784.

In particular, this change allowed me to build the Zephyr
RTOS for riscv with `CONFIG_NEWLIB_LIBC=y` without any
such linker errors.

Fixes #57

Signed-off-by: Christopher Friedt <chrisfriedt@gmail.com>